+++ src/commonMain/kotlin/entities/ProblemDetails.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable internal data class ProblemDetails(
     val type: String? = null,
     val title: String? = null,
-    val status: Int? = null,
+    val status: Short? = null,
     val detail: String? = null,
     val instance: String? = null,
 )
