+++ src/commonMain/kotlin/entities/WorldItemRecency.kt
@@ -13,6 +13,6 @@ import kotlinx.serialization.Serializable
 @JsExport @Serializable data class WorldItemRecency(
     @SerialName("itemID") val itemId: Int,
     val lastUploadTime: Long,
-    @SerialName("worldID") val worldId: Int,
+    @SerialName("worldID") val worldId: Short,
     val worldName: String? = null,
 )
